Save the username of a user
The username is saved when he succesfully enters a game, either single
or multiplayer (not waiting room).
The username is saved to `username.txt`, which is created if it doesn't
exist yet. This file has also been added to .gitignore
When the user enters the splash screen the saved username will pe
autofilled in the usernameField, but it is not checked for duplication
until the user enters a game.